restyle, address code review
