+++ components/icons/NavbarIcons.tsx
@@ -112,9 +112,9 @@ export const MenuLogo: React.FC = ({}) => (
     <path
       d="M2.25 21H21.75M3.75 3V21M14.25 3V21M20.25 7.5V21M6.75 6.75H7.5M6.75 9.75H7.5M6.75 12.75H7.5M10.5 6.75H11.25M10.5 9.75H11.25M10.5 12.75H11.25M6.75 21V17.625C6.75 17.0037 7.25368 16.5 7.875 16.5H10.125C10.7463 16.5 11.25 17.0037 11.25 17.625V21M3 3H15M14.25 7.5H21M17.25 11.25H17.2575V11.2575H17.25V11.25ZM17.25 14.25H17.2575V14.2575H17.25V14.25ZM17.25 17.25H17.2575V17.2575H17.25V17.25Z"
       stroke="white"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
+      strokeWidth="1.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
     />
     <path
       d="M48.3395 8.45312C48.2732 7.83286 47.9938 7.34991 47.5014 7.00426C47.0137 6.65862 46.3793 6.4858 45.598 6.4858C45.0488 6.4858 44.5777 6.56866 44.1847 6.73438C43.7917 6.90009 43.491 7.125 43.2827 7.40909C43.0743 7.69318 42.9678 8.01752 42.9631 8.3821C42.9631 8.68513 43.0317 8.94792 43.169 9.17045C43.3111 9.39299 43.5028 9.58239 43.7443 9.73864C43.9858 9.89015 44.2533 10.018 44.5469 10.1222C44.8404 10.2263 45.1364 10.3139 45.4347 10.3849L46.7983 10.7259C47.3475 10.8537 47.8755 11.0265 48.3821 11.2443C48.8935 11.4621 49.3504 11.7367 49.7528 12.0682C50.16 12.3996 50.482 12.7997 50.7188 13.2685C50.9555 13.7372 51.0739 14.2865 51.0739 14.9162C51.0739 15.7685 50.8561 16.5189 50.4205 17.1676C49.9848 17.8116 49.3551 18.3158 48.5312 18.6804C47.7121 19.0402 46.7202 19.2202 45.5554 19.2202C44.4238 19.2202 43.4413 19.045 42.608 18.6946C41.7794 18.3442 41.1307 17.8329 40.6619 17.1605C40.1979 16.4882 39.947 15.669 39.9091 14.7031H42.5014C42.5393 15.2098 42.6955 15.6312 42.9702 15.9673C43.2448 16.3035 43.6023 16.5545 44.0426 16.7202C44.4877 16.8859 44.9848 16.9688 45.5341 16.9688C46.107 16.9688 46.6089 16.8835 47.0398 16.7131C47.4754 16.5379 47.8163 16.2964 48.0625 15.9886C48.3087 15.6761 48.4342 15.3116 48.4389 14.8949C48.4342 14.5161 48.3229 14.2036 48.1051 13.9574C47.8873 13.7064 47.5819 13.4981 47.1889 13.3324C46.8007 13.1619 46.3461 13.0104 45.8253 12.8778L44.1705 12.4517C42.9725 12.1439 42.0256 11.6776 41.3295 11.0526C40.6383 10.4228 40.2926 9.58712 40.2926 8.54545C40.2926 7.68845 40.5246 6.93797 40.9886 6.29403C41.4574 5.65009 42.0942 5.15057 42.8991 4.79545C43.7041 4.43561 44.6155 4.25568 45.6335 4.25568C46.6657 4.25568 47.5701 4.43561 48.3466 4.79545C49.1278 5.15057 49.741 5.64536 50.1861 6.27983C50.6312 6.90956 50.8608 7.634 50.875 8.45312H48.3395ZM53.3093 19V8.09091H55.8803V19H53.3093ZM54.6019 6.54261C54.1947 6.54261 53.8443 6.40767 53.5508 6.13778C53.2572 5.86316 53.1104 5.53409 53.1104 5.15057C53.1104 4.76231 53.2572 4.43324 53.5508 4.16335C53.8443 3.88873 54.1947 3.75142 54.6019 3.75142C55.0138 3.75142 55.3642 3.88873 55.6531 4.16335C55.9466 4.43324 56.0934 4.76231 56.0934 5.15057C56.0934 5.53409 55.9466 5.86316 55.6531 6.13778C55.3642 6.40767 55.0138 6.54261 54.6019 6.54261ZM63.9716 8.09091V10.0795H57.7003V8.09091H63.9716ZM59.2486 5.47727H61.8196V15.7188C61.8196 16.0644 61.8717 16.3295 61.9759 16.5142C62.0848 16.6941 62.2268 16.8172 62.402 16.8835C62.5772 16.9498 62.7713 16.983 62.9844 16.983C63.1454 16.983 63.2921 16.9711 63.4247 16.9474C63.562 16.9238 63.6662 16.9025 63.7372 16.8835L64.1705 18.8935C64.0331 18.9408 63.8366 18.9929 63.581 19.0497C63.33 19.1065 63.0223 19.1397 62.6577 19.1491C62.0137 19.1681 61.4337 19.071 60.9176 18.858C60.4015 18.6402 59.992 18.304 59.6889 17.8494C59.3906 17.3949 59.2438 16.8267 59.2486 16.1449V5.47727ZM70.8342 19.2131C69.7404 19.2131 68.7958 18.9858 68.0004 18.5312C67.2096 18.072 66.6012 17.4233 66.1751 16.5852C65.7489 15.7424 65.5359 14.7505 65.5359 13.6094C65.5359 12.4872 65.7489 11.5024 66.1751 10.6548C66.6059 9.80256 67.2073 9.13968 67.979 8.66619C68.7508 8.18797 69.6576 7.94886 70.6992 7.94886C71.3716 7.94886 72.006 8.05777 72.6026 8.27557C73.204 8.48864 73.7343 8.82008 74.1935 9.26989C74.6576 9.7197 75.0221 10.2926 75.2873 10.9886C75.5524 11.6799 75.685 12.5038 75.685 13.4602V14.2486H66.7433V12.5156H73.2205C73.2158 12.0232 73.1093 11.5852 72.9009 11.2017C72.6926 10.8134 72.4014 10.508 72.0273 10.2855C71.658 10.063 71.2272 9.9517 70.7347 9.9517C70.2092 9.9517 69.7475 10.0795 69.3498 10.3352C68.9521 10.5862 68.6419 10.9176 68.4194 11.3295C68.2016 11.7367 68.0903 12.1842 68.0856 12.6719V14.1847C68.0856 14.8191 68.2016 15.3636 68.4336 15.8182C68.6656 16.268 68.9899 16.6136 69.4066 16.8551C69.8233 17.0919 70.311 17.2102 70.8697 17.2102C71.2437 17.2102 71.5823 17.1581 71.8853 17.054C72.1883 16.9451 72.4511 16.7865 72.6737 16.5781C72.8962 16.3698 73.0643 16.1117 73.1779 15.804L75.5785 16.0739C75.427 16.7083 75.1381 17.2623 74.712 17.7358C74.2906 18.2045 73.7508 18.5691 73.0927 18.8295C72.4345 19.0852 71.6817 19.2131 70.8342 19.2131ZM86.4822 10.9744L84.1385 11.2301C84.0722 10.9934 83.9562 10.7708 83.7905 10.5625C83.6295 10.3542 83.4117 10.1861 83.1371 10.0582C82.8625 9.9304 82.5263 9.86648 82.1286 9.86648C81.5935 9.86648 81.1437 9.98248 80.7791 10.2145C80.4193 10.4465 80.2417 10.7472 80.2464 11.1165C80.2417 11.4337 80.3577 11.6918 80.5945 11.8906C80.8359 12.0895 81.2337 12.2528 81.7876 12.3807L83.6484 12.7784C84.6806 13.0009 85.4477 13.3537 85.9496 13.8366C86.4562 14.3196 86.7119 14.9517 86.7166 15.733C86.7119 16.4195 86.5107 17.0256 86.1129 17.5511C85.7199 18.072 85.1731 18.4792 84.4723 18.7727C83.7715 19.0663 82.9666 19.2131 82.0575 19.2131C80.7223 19.2131 79.6475 18.9337 78.8331 18.375C78.0187 17.8116 77.5334 17.0279 77.3771 16.0241L79.8842 15.7827C79.9979 16.2751 80.2393 16.6468 80.6087 16.8977C80.978 17.1487 81.4586 17.2741 82.0504 17.2741C82.6612 17.2741 83.1513 17.1487 83.5206 16.8977C83.8946 16.6468 84.0817 16.3366 84.0817 15.9673C84.0817 15.6548 83.9609 15.3968 83.7195 15.1932C83.4827 14.9896 83.1134 14.8333 82.6115 14.7244L80.7507 14.3338C79.7043 14.116 78.9302 13.7491 78.4283 13.233C77.9264 12.7121 77.6778 12.054 77.6825 11.2585C77.6778 10.5862 77.8601 10.0038 78.2294 9.51136C78.6035 9.0142 79.1219 8.63068 79.7848 8.3608C80.4524 8.08617 81.2218 7.94886 82.093 7.94886C83.3714 7.94886 84.3776 8.22112 85.1115 8.76562C85.8501 9.31013 86.3071 10.0464 86.4822 10.9744Z"
